Backend/requirements.txt

azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-cognitiveservices-nspkg==3.0.1
azure-cognitiveservices-search-imagesearch==2.0.0
azure-cognitiveservices-search-nspkg==3.0.1
azure-cognitiveservices-search-websearch==1.0.0
azure-mgmt-cognitiveservices==3.0.0

bs4==0.0.1
certifi==2019.9.11
Flask==1.1.1
Flask-Cors==3.0.8
httplib2==0.9.2
nltk==3.4.5
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
scipy==1.3.1
virtualenv==16.7.2
